# Add GPU mode for resize transform