--- a/chainercv/utils/bbox/non_maximum_suppression.py
+++ b/chainercv/utils/bbox/non_maximum_suppression.py
@@ -7,12 +7,12 @@ from chainercv.utils.bbox._nms_gpu_post import _nms_gpu_post
 
 
 if cuda.available:
-    import cupy
+    import cupy as cp
 
-    @cupy.util.memoize(for_each_device=True)
+    @cp.util.memoize(for_each_device=True)
     def _load_kernel(kernel_name, code, options=()):
         assert isinstance(options, tuple)
-        kernel_code = cupy.cuda.compile_with_cache(code, options=options)
+        kernel_code = cp.cuda.compile_with_cache(code, options=options)
         return kernel_code.get_function(kernel_name)
 
 
@@ -38,7 +38,7 @@ def non_maximum_suppression(bbox, thresh, score=None,
     :obj:`score` is a float array of shape :math:`(R,)`. Each score indicates
     confidence of prediction.
 
-    This function accepts both :obj:`numpy.ndarray` and :obj:`cupy.ndarray` as
+    This function accepts both :obj:`numpy.ndarray` and :obj:`cp.ndarray` as
     inputs. Please note that both :obj:`bbox` and :obj:`score` need to be
     same type.
     The output is same type as the type of the inputs.
@@ -100,16 +100,16 @@ def _non_maximum_suppression_cpu(bbox, thresh, score=None, limit=None):
 
 def _non_maximum_suppression_gpu(bbox, thresh, score=None, limit=None):
     if len(bbox) == 0:
-        return cupy.zeros((0,), dtype=np.int32)
+        return cp.zeros((0,), dtype=np.int32)
 
     n_bbox = bbox.shape[0]
 
     if score is not None:
-        # CuPy does not currently support argsort.
+        # cp does not currently support argsort.
         order = cuda.to_cpu(score).argsort()[::-1].astype(np.int32)
         order = cuda.to_gpu(order)
     else:
-        order = cupy.arange(n_bbox, dtype=np.int32)
+        order = cp.arange(n_bbox, dtype=np.int32)
 
     sorted_bbox = bbox[order, :]
     selec, n_selec = _call_nms_kernel(
@@ -193,10 +193,10 @@ def _call_nms_kernel(bbox, thresh):
     blocks = (col_blocks, col_blocks, 1)
     threads = (threads_per_block, 1, 1)
 
-    mask_dev = cupy.zeros((n_bbox * col_blocks,), dtype=np.uint64)
-    bbox = cupy.ascontiguousarray(bbox, dtype=np.float32)
+    mask_dev = cp.zeros((n_bbox * col_blocks,), dtype=np.uint64)
+    bbox = cp.ascontiguousarray(bbox, dtype=np.float32)
     kern = _load_kernel('nms_kernel', _nms_gpu_code)
-    kern(blocks, threads, args=(n_bbox, cupy.float32(thresh),
+    kern(blocks, threads, args=(n_bbox, cp.float32(thresh),
                                 bbox, mask_dev))
 
     mask_host = mask_dev.get()