### Demo

```python
from chainercv.transforms import resize
import numpy as np
import chainer
from skimage.data import astronaut
import matplotlib.pyplot as plt
from chainercv.visualizations import vis_image

src = chainer.cuda.to_gpu(astronaut().transpose(2, 0, 1).astype(np.float32))
size = tuple(np.random.randint(224, 480, size=(2,)))
dst = resize(src, size)
vis_image(dst.get())
plt.show()
```


### benchmark

```python
from chainercv.transforms import resize
import numpy as np
import chainer
from skimage.data import astronaut


src = chainer.cuda.to_gpu(astronaut().transpose(2, 0, 1).astype(np.float32))
src_cpu = src.get()
size = (224, 280)

def benchmark(func, args, n_run):
    import cupy as cp
    times = []
    for _ in range(n_run):
        start = cp.cuda.Event()
        end = cp.cuda.Event()
        start.record()
        func(*args)
        end.record()
        end.synchronize()
        times.append(cp.cuda.get_elapsed_time(start, end))  # milliseconds
    return times

# dry runs
for _ in range(3):
    dst = resize(src, size)

print('gpu {}ms'.format(np.mean(benchmark(resize, (src, size), 10))))


import time
start = time.time()
resize(src_cpu, size)
print('cpu {}ms'.format((time.time() - start) * 1000))
```

#### benchmark result
```
gpu 0.0970143988729ms 
cpu 28.1879901886ms   
```