@Hakuyume 2017-07-03
How about making another PR for `cupy -> cp`?