@yuyu2172 2017-07-05
I changed `resize_gpu` to launch a thread for each channel.
On my setting, the kernel execution time improved slightly.
```
Previously (Quadro M1000M)
 42.21%  725.38us        13  55.798us  52.833us  65.889us  resize_linear
New (Quadro M1000M)
 34.86%  710.09us        13  54.622us  52.960us  58.497us  resize_linear

Previous (TitanX)
 10.98%  133.99us        13  10.306us  9.2800us  12.768us  resize_linear
New (TitanX)
 10.30%  127.14us        13  9.7800us  8.9610us  11.744us  resize_linear

```
I used the benchmark in the previous comments.