@yuyu2172 2017-11-29
I will temporarily close this because we do not have a plan to include this in ChainerCV.